backend/requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
websockets>=11.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Data Processing
albumentations>=1.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pandas>=2.0.0

# System & Utilities
psutil>=5.9.0
pyyaml>=6.0.0
rasterio>=1.3.0
requests>=2.31.0
huggingface-hub>=0.16.0

# Security
cryptography>=41.0.0

# OCR (optional - install if using OCR features)
# easyocr>=1.7.0
# pytesseract>=0.3.10
